Move backend api/model classes to their own module
Move backend api/model classes from model to backends/spi module as
those are not required in order to use/implement the Nessie REST api.
