+++ backends/dynamodb/pom.xml
@@ -32,7 +32,7 @@
   <dependencies>
     <dependency>
       <groupId>org.projectnessie</groupId>
-      <artifactId>nessie-model</artifactId>
+      <artifactId>nessie-backends-spi</artifactId>
       <version>${project.version}</version>
     </dependency>
     <dependency>
